requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
docker==5.0.0
fast-json==0.3.2
idna==3.2
paramiko==2.7.2
pkg_resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
sshtunnel==0.4.0
ujson==4.1.0
urllib3==1.26.6
websocket-client==1.2.1
